+++ packages/coauthors-bot/package.json
@@ -22,8 +22,11 @@
   },
   "sideEffects": false,
   "type": "module",
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "build": "tsc",
+    "build": "tsup",
     "ci:test": "vitest run --coverage --typecheck",
     "ci:test:watch": "vitest --ui --coverage --typecheck",
     "dev": "nodemon",
@@ -36,6 +39,7 @@
   "devDependencies": {
     "@coauthors/eslint-config-ts": "workspace:*",
     "@coauthors/tsconfig": "workspace:*",
+    "@coauthors/tsup": "workspace:*",
     "nock": "^14.0.0-beta.5",
     "nodemon": "^3.1.4",
     "smee-client": "^2.0.0"
