+++ packages/coauthors-bot/tsconfig.json
@@ -1,26 +1,7 @@
 {
+  "extends": "@coauthors/tsconfig/base.json",
+  "include": [".", "./.eslintrc.cjs"],
   "compilerOptions": {
-    "incremental": true,
-    "target": "es2022",
-    "module": "Node16",
-    "declaration": true,
-    "sourceMap": true,
-    "outDir": "./dist",
-
-    /* Strict Type-Checking Options */
-    "strict": true,
-
-    /* Additional Checks */
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-
-    "moduleResolution": "Node16",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true
-  },
-  "include": ["src/", "./.eslintrc.cjs"],
-  "exclude": ["node_modules", "*.test.*", "*.spec.*", "*.test-d.*"],
-  "compileOnSave": false
+    "types": ["vitest/globals"]
+  }
 }
