+++ packages/coauthors-bot/tsup.config.ts
@@ -1,0 +1,4 @@
+import { options } from '@coauthors/tsup'
+import { defineConfig } from 'tsup'
+
+export default defineConfig(options)
